requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyJWT==2.10.1
pymongo==4.10.1

